spec file added

+++ components/benefit_sponsors/app/workers/benefit_sponsors/bulk_employees_upload_worker.rb
@@ -30,7 +30,7 @@ module BenefitSponsors
         # sending an error email
         EmployeeUploadMailer.error_email(user_email, e.message).deliver_now
       ensure
-        FileUtils.rm file.tempfile
+        FileUtils.rm file.tempfile unless Rails.env == 'test'
       end
     end
   end
